a_onsite_lectures/class10-20240114/todo_api_pytest/requirements.txt
python-dotenv
SQLAlchemy
fastapi
uvicorn
pydantic
requests